% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqMetaPipeline.R
\docType{data}
\name{pheno1}
\alias{pheno1}
\title{Phenotype data for 5700 subjects}
\format{A data frame with 5718 rows and 5 variables:
\describe{
  \item{id}{subject identifier}
  \item{y}{phenotype measure}
  \item{age}{age of the subject}
  \item{pc1}{principle component 1}
  \item{pc2}{principle component 2}
}}
\source{
Internal randomized dataset
}
\usage{
pheno1
}
\description{
A dataset containing the phenotype data for 5700 subjects
}
\keyword{datasets}

