% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqMetaPipeline.R
\docType{package}
\name{seqMetaPipeline}
\alias{seqMetaPipeline}
\alias{seqMetaPipeline-package}
\title{seqMetaPipeline: Analysis Pipeline for SeqMeta}
\description{
Functions to make seqMeta analyses and reports consistent.
}
\details{
To learn more about seqMetaPipeline, start with the vignettes:
}

